# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ApplyLeaveRequest:
    properties:
      end_date:
        example: "2025-12-05"
        type: string
      leave_type_id:
        example: 1
        type: integer
      reason:
        example: Family vacation
        type: string
      start_date:
        example: "2025-12-01"
        type: string
    required:
    - end_date
    - leave_type_id
    - start_date
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        example: Invalid credentials
        type: string
    type: object
...
  handlers.LeaveBalanceResponse:
    properties:
      balance:
        example: 15
        type: integer
      leave_type_id:
        example: 1
        type: integer
      leave_type_name:
        example: Annual
        type: string
      max_days:
        example: 20
        type: integer
      used_days:
        example: 5
        type: integer
    type: object
...
  handlers.RejectLeaveRequest:
    properties:
      reason:
        example: Insufficient staffing during requested period
        type: string
    required:
    - reason
    type: object
...
  models.AuditAction:
    enum:
    - CREATE
    - APPROVE
    - REJECT
    - CANCEL
    - UPDATE
    - DELETE
    type: string
    x-enum-varnames:
    - AuditActionCreate
    - AuditActionApprove
    - AuditActionReject
    - AuditActionCancel
    - AuditActionUpdate
    - AuditActionDelete
...
  models.AuditLog:
    properties:
      action:
        $ref: '#/definitions/models.AuditAction'
      changes:
        description: JSON representation of what changed
        type: string
      comment:
        type: string
      created_at:
        type: string
      entity_id:
        type: integer
      entity_type:
        $ref: '#/definitions/models.AuditEntityType'
      id:
        type: integer
      ip_address:
        type: string
      new_values:
        description: JSON representation of new values
        type: string
      old_values:
        description: JSON representation of old values
        type: string
      performed_by:
        type: integer
      performer:
        $ref: '#/definitions/models.Employee'
      request_method:
        type: string
      request_path:
        type: string
      user_agent:
        type: string
    type: object
  models.ComplianceRecord:
    properties:
      created_at:
        type: string
      document:
        $ref: '#/definitions/models.Document'
      document_id:
        type: integer
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      expiry_date:
        type: string
      id:
        type: integer
      issue_date:
        type: string
      last_verified_date:
        type: string
      non_compliance_reason:
        type: string
      notes:
        type: string
      requirement:
        $ref: '#/definitions/models.ComplianceRequirement'
      requirement_id:
        type: integer
      status:
        $ref: '#/definitions/models.ComplianceStatus'
      updated_at:
        type: string
      verified_by:
        type: integer
      verifier:
        $ref: '#/definitions/models.Employee'
    type: object
...
  models.Document:
    properties:
      created_at:
        type: string
      description:
        type: string
      document_type:
        $ref: '#/definitions/models.DocumentType'
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      expiry_date:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      id:
        type: integer
      is_confidential:
        type: boolean
      issue_date:
        type: string
      mime_type:
        type: string
      status:
        $ref: '#/definitions/models.DocumentStatus'
      tags:
        type: string
      title:
        type: string
      updated_at:
        type: string
      uploaded_by:
        type: integer
      uploader:
        $ref: '#/definitions/models.Employee'
      verified_at:
        type: string
      verified_by:
        type: integer
      verifier:
        $ref: '#/definitions/models.Employee'
    type: object
...
  models.Employee:
    properties:
      audit_logs:
        items:
          $ref: '#/definitions/models.AuditLog'
        type: array
      compliance_records:
        items:
          $ref: '#/definitions/models.ComplianceRecord'
        type: array
      created_at:
        type: string
      department:
        type: string
      documents:
        items:
          $ref: '#/definitions/models.Document'
        type: array
      email:
        type: string
      employment:
        $ref: '#/definitions/models.EmploymentDetails'
      employment_history:
        items:
          $ref: '#/definitions/models.EmploymentHistory'
        type: array
      firstname:
        type: string
      id:
        type: integer
      identity:
        $ref: '#/definitions/models.IdentityInformation'
      lastname:
        type: string
      leaves:
        description: Relationships
        items:
          $ref: '#/definitions/models.Leave'
        type: array
      lifecycle_events:
        items:
          $ref: '#/definitions/models.WorkLifecycleEvent'
        type: array
      nrc:
        type: string
      offboarding_process:
        $ref: '#/definitions/models.OffboardingProcess'
      onboarding_process:
        $ref: '#/definitions/models.OnboardingProcess'
      position:
        $ref: '#/definitions/models.Position'
      position_assignments:
        items:
          $ref: '#/definitions/models.PositionAssignment'
        type: array
      position_id:
        type: integer
      role:
        $ref: '#/definitions/models.Role'
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.EmploymentDetails:
    properties:
      created_at:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      employee_number:
        type: string
      employment_status:
        $ref: '#/definitions/models.EmploymentStatus'
      employment_type:
        $ref: '#/definitions/models.EmploymentType'
      end_date:
        type: string
      hire_date:
        type: string
      id:
        type: integer
      manager:
        $ref: '#/definitions/models.Employee'
      manager_id:
        type: integer
      notice_period:
        description: in days
        type: integer
      probation_end_date:
        type: string
      probation_status:
        type: string
      start_date:
        type: string
      termination_date:
        type: string
      termination_reason:
        type: string
      updated_at:
        type: string
      work_location:
        type: string
      work_schedule:
        type: string
    type: object
  models.EmploymentHistory:
    properties:
      change_date:
        type: string
      change_reason:
        type: string
      changed_by:
        type: integer
      changer:
        $ref: '#/definitions/models.Employee'
      created_at:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      id:
        type: integer
      new_department:
        type: string
      new_position:
        type: string
      new_status:
        $ref: '#/definitions/models.EmploymentStatus'
      notes:
        type: string
      previous_department:
        type: string
      previous_position:
        type: string
      previous_status:
        $ref: '#/definitions/models.EmploymentStatus'
      updated_at:
        type: string
    type: object
...
  models.IdentityInformation:
    properties:
      address:
        type: string
      blood_group:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      emergency_contact:
        type: string
      emergency_phone:
        type: string
      emergency_relation:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      gender:
        type: string
      id:
        type: integer
      marital_status:
        type: string
      mobile_number:
        type: string
      nationality:
        type: string
      phone_number:
        type: string
      postal_code:
        type: string
      state:
        type: string
      updated_at:
        type: string
    type: object
  models.Leave:
    properties:
      approved_at:
        type: string
      approved_by:
        type: integer
      approver:
        $ref: '#/definitions/models.Employee'
      created_at:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      end_date:
        type: string
      id:
        type: integer
      leave_type:
        $ref: '#/definitions/models.LeaveType'
      leave_type_id:
        type: integer
      reason:
        type: string
      rejection_reason:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/models.LeaveStatus'
      updated_at:
        type: string
    type: object
...
  models.LeaveAudit:
    properties:
      action:
        $ref: '#/definitions/models.AuditAction'
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      leave:
        $ref: '#/definitions/models.Leave'
      leave_id:
        type: integer
      new_status:
        type: string
      old_status:
        type: string
      performed_by:
        type: integer
      performer:
        $ref: '#/definitions/models.Employee'
    type: object
  models.LeaveStatus:
    enum:
    - Pending
    - Approved
    - Rejected
    - Cancelled
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusApproved
    - StatusRejected
    - StatusCancelled
  models.LeaveType:
    properties:
      created_at:
        type: string
      id:
        type: integer
      leaves:
        items:
          $ref: '#/definitions/models.Leave'
        type: array
      max_days:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.OffboardingProcess:
    properties:
      actual_end_date:
        type: string
      assigned_to:
        type: integer
      assignee:
        $ref: '#/definitions/models.Employee'
      created_at:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      expected_end_date:
        type: string
      id:
        type: integer
      initiated_by:
        type: integer
      initiator:
        $ref: '#/definitions/models.Employee'
      notes:
        type: string
      reason:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/models.OnboardingStatus'
      tasks:
        items:
          $ref: '#/definitions/models.OffboardingTask'
        type: array
      updated_at:
        type: string
    type: object
...
  models.OnboardingProcess:
    properties:
      actual_end_date:
        type: string
      assigned_to:
        type: integer
      assignee:
        $ref: '#/definitions/models.Employee'
      created_at:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      expected_end_date:
        type: string
      id:
        type: integer
      initiated_by:
        type: integer
      initiator:
        $ref: '#/definitions/models.Employee'
      notes:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/models.OnboardingStatus'
      tasks:
        items:
          $ref: '#/definitions/models.OnboardingTask'
        type: array
      updated_at:
        type: string
    type: object
...
  models.WorkLifecycleEvent:
    properties:
      approved_at:
        type: string
      approved_by:
        type: integer
      approver:
        $ref: '#/definitions/models.Employee'
      completion_date:
        type: string
      created_at:
        type: string
      description:
        type: string
      effective_date:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      event_date:
        type: string
      event_type:
        $ref: '#/definitions/models.LifecycleEventType'
      id:
        type: integer
      initiated_by:
        type: integer
      initiator:
        $ref: '#/definitions/models.Employee'
      is_completed:
        type: boolean
      new_value:
        type: string
      notes:
        type: string
      previous_value:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8070
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    REST API for Leave Management System - part of HRMS
    Allows employees to apply for leaves, track balances, and managers to approve/reject requests.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: HRMS Leave Management API
  version: "1.0"
paths:
...
  /api/leaves:
    get:
      description: Get all leave requests for the authenticated employee
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Leave'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my leave history
      tags:
      - Leaves
    post:
      consumes:
      - application/json
      description: Submit a new leave request
      parameters:
      - description: Leave application data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ApplyLeaveRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Leave'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Overlapping leave exists
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Apply for leave
      tags:
      - Leaves
  /api/leaves/{id}/approve:
    put:
      description: Approve a pending leave request (Manager/Admin only)
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Leave'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Approve leave
      tags:
      - Manager
  /api/leaves/{id}/audit:
    get:
      description: Get audit history for a leave request (Manager/Admin only)
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LeaveAudit'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get leave audit trail
      tags:
      - Manager
  /api/leaves/{id}/cancel:
    put:
      description: Cancel own pending or approved leave request
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Leave'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel leave
      tags:
      - Leaves
  /api/leaves/{id}/reject:
    put:
      consumes:
      - application/json
      description: Reject a pending leave request with reason (Manager/Admin only)
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: integer
      - description: Rejection reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RejectLeaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Leave'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reject leave
      tags:
      - Manager
  /api/leaves/balance:
    get:
      description: Get remaining leave balance for all leave types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.LeaveBalanceResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get leave balance
      tags:
      - Leaves
  /api/leaves/pending:
    get:
      description: Get all pending leave requests (Manager/Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Leave'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get pending leaves
      tags:
      - Manager
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" (without quotes)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"